authService/requirements.txt
DJango>=3.0.6,<3.1
wheel
djangorestframework>=3.12.2,<3.13.0
psycopg2-binary==2.8.6
dj-rest-auth>=2.1,<2.2
djangorestframework-simplejwt>=4.6.0,<4.7
django-allauth>=0.44.0,<0.45
django-countries
django-filter
django-rest-auth